move tool registration to different file
